labs/analytics/emr-analytics/requirements.txt
aws-cdk-lib==2.52.0
constructs>=10.0.0,<11.0.0
